# backend/docs/swagger.yaml
basePath: /
definitions:
  web-hosting_internal_database_entities.Jurusan:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  web-hosting_internal_modules_prodi_dto.ProdiCreateRequest:
    properties:
      jenjang:
        enum:
        - D3
        - D4
        example: D3
        type: string
      jurusan_name:
        example: teknik-elektro
        type: string
      name:
        example: teknik-elektro
        type: string
    required:
    - jenjang
    - jurusan_name
    - name
    type: object
  web-hosting_internal_modules_prodi_dto.ProdiResponse:
    properties:
      id:
        example: 1
        type: integer
      jenjang:
        example: D3
        type: string
      jurusan:
        $ref: '#/definitions/web-hosting_internal_database_entities.Jurusan'
      name:
        example: teknik-elektro
        type: string
    type: object
  web-hosting_internal_modules_prodi_dto.ProdiUpdateRequest:
    properties:
      jenjang:
        example: D3
        type: string
      jurusan_id:
        example: 1
        type: integer
      name:
        example: teknik-elektro
        type: string
    type: object
...
  web-hosting_internal_package_swagger.ErrCreateProdiFailed:
    properties:
      error:
        example: prodi already exists
        type: string
      message:
        example: failed to create prodi
        type: string
      path:
        example: /api/prodi
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrCreateProdiInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to create prodi
        type: string
      path:
        example: /api/prodi
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrDeleteProdiFailed:
    properties:
      error:
        example: prodi already exists
        type: string
      message:
        example: failed to Delete prodi
        type: string
      path:
        example: /api/prodi
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrDeleteProdiInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to Delete prodi
        type: string
      path:
        example: /api/prodi
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrForbiddenAccess:
    properties:
      error:
        example: Forbidden
        type: string
      message:
        example: Role anda tidak diizinkan
        type: string
      path:
        example: /api/super/role
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrGetProdiFailed:
    properties:
      error:
        example: prodi not found
        type: string
      message:
        example: failed to get prodi
        type: string
      path:
        example: /api/prodi
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrGetProdiInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to get prodi
        type: string
      path:
        example: /api/prodi
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken:
    properties:
      error:
        example: invalid token
        type: string
      message:
        example: failed_auth
        type: string
      path:
        example: /api/auth/logout
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_swagger.ErrUpdateProdiFailed:
    properties:
      error:
        example: prodi already exists
        type: string
      message:
        example: failed to update prodi
        type: string
      path:
        example: /api/prodi
        type: string
      success:
        example: false
        type: boolean
    type: object
  web-hosting_internal_package_swagger.ErrUpdateProdiInternalServer:
    properties:
      error:
        example: Internal Error
        type: string
      message:
        example: failed to update prodi
        type: string
      path:
        example: /api/prodi
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  web-hosting_internal_package_utils.Response-any-any:
    properties:
      data: {}
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
  web-hosting_internal_package_utils.Response-array_web-hosting_internal_modules_prodi_dto_ProdiResponse-any:
    properties:
      data:
        items:
          $ref: '#/definitions/web-hosting_internal_modules_prodi_dto.ProdiResponse'
        type: array
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
  web-hosting_internal_package_utils.Response-web-hosting_internal_modules_prodi_dto_ProdiResponse-any:
    properties:
      data:
        $ref: '#/definitions/web-hosting_internal_modules_prodi_dto.ProdiResponse'
      error: {}
      message:
        example: Operation successful
        type: string
      path:
        example: /api/resource
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    **BARANG SIAPA YANG UPDATE/HAPUS DATA SEMBARANGAN DAPAT KECUPAN DARI HAFIZ**.
    **BUAT DATA BARU JIKA INGIN HAPUS/UPDATE**.

    **DUMMY DATA LINK:**
    **https://github.com/Kar-Su/pbl-project-simpadu/tree/main/backend/internal/database/seeders/json**

    **TUTORIAL LINK:***
    **https://github.com/Kar-Su/pbl-project-simpadu**
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TIM 1 API
  version: "1.0"
paths:
...
  /api/prodi:
    get:
      consumes:
      - application/json
      description: |-
        Mendapatkan data prodi. Jika query id/name diberikan, return 1 data spesifik; tanpa query, return semua data

        **Akses:** Logged User
      parameters:
      - description: Nama Prodi
        example: teknik-elektro
        in: query
        name: name
        type: string
      - description: ID Prodi
        example: 1
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-array_web-hosting_internal_modules_prodi_dto_ProdiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetProdiFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUpdateProdiInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Get Prodi
      tags:
      - prodi
    post:
      consumes:
      - application/json
      description: |-
        Menambahkan prodi baru ke sistem

        **Akses:** Admin Akademik
      parameters:
      - description: Prodi Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web-hosting_internal_modules_prodi_dto.ProdiCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-any-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrCreateProdiFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrCreateProdiInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Create Prodi Baru
      tags:
      - prodi
  /api/prodi/{name}:
    delete:
      consumes:
      - application/json
      description: |-
        Menghapus prodi berdasarkan nama di path parameter

        **Akses:** Admin Akademik
      parameters:
      - description: Nama Prodi yang akan dihapus
        example: teknik-elektro
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-any-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrDeleteProdiFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrDeleteProdiInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Delete Prodi
      tags:
      - prodi
    put:
      consumes:
      - application/json
      description: |-
        Mengupdate prodi berdasarkan nama di path parameter

        **Akses:** Admin Akademik
      parameters:
      - description: Nama Prodi yang akan diupdate
        example: teknik-elektro
        in: path
        name: name
        required: true
        type: string
      - description: Prodi Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web-hosting_internal_modules_prodi_dto.ProdiUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-web-hosting_internal_modules_prodi_dto_ProdiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUpdateProdiFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrForbiddenAccess'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUpdateProdiInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Update Prodi
      tags:
      - prodi
  /api/prodi/jurusan/{jurusan_name}:
    get:
      consumes:
      - application/json
      description: |-
        Mendapatkan daftar prodi berdasarkan nama jurusan

        **Akses:** Logged User
      parameters:
      - description: Nama Jurusan
        example: teknik-elektro
        in: path
        name: jurusan_name
        required: true
        type: string
      - description: 'Halaman (default: 1, per halaman: 10)'
        example: 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web-hosting_internal_package_utils.Response-array_web-hosting_internal_modules_prodi_dto_ProdiResponse-any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetProdiFailed'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrUnauthorizedInvalidToken'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web-hosting_internal_package_swagger.ErrGetProdiInternalServer'
      security:
      - ApiKeyAuth: []
      summary: Get Prodi by Jurusan
      tags:
      - prodi
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Format: Bearer <access token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"